better startup error handling for celanup

+++ src/quorum/daemon.py
@@ -66,6 +66,10 @@ class Daemon:
     """ The file path to the file to be used
     as the standard error of the created process """
 
+    error = False
+    """ The flag indicating if a startup error as ocured in
+    the current daemon, avoid problems with pid """
+
     def __init__(self, pid_file, stdin = "/dev/null", stdout = "/dev/null", stderr = "/dev/null"):
         """
         Constructor of the class.
@@ -157,6 +161,7 @@ class Daemon:
         # process (avoids duplicated running)
         if pid:
             message = "pidfile %s already exists, daemon already running ?\n"
+            self.error = True
             sys.stderr.write(message % self.pidfile)
             sys.exit(1)
 
@@ -192,7 +197,6 @@ class Daemon:
                 pid_exists = os.path.exists(self.pidfile)
                 pid_exists and os.remove(self.pidfile)
             else:
-                print str(err)
                 sys.exit(1)
 
     def restart(self):
@@ -210,6 +214,7 @@ class Daemon:
         releasing all the structures locked by it.
         """
 
+        if self.error: return
         pid_exists = os.path.exists(self.pidfile)
         pid_exists and os.remove(self.pidfile)
